ultima actualizacion del sistema 19/08/2025

+++ src/thebarbershop/Jframe/PerfilCliente.java
@@ -280,7 +280,22 @@ public class PerfilCliente extends javax.swing.JFrame {
     }//GEN-LAST:event_JBguardarActionPerformed
 
     private void JBeliminarPerfilActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_JBeliminarPerfilActionPerformed
+        int confirmacion = JOptionPane.showConfirmDialog(
+        this,
+        "<html><b>¿Estás seguro de que deseas eliminar tu cuenta?</b><br><br>" +
+        "Esta acción es <b>irreversible</b>.<br>" +
+        "Se eliminarán todos tus datos y no podrás recuperar tu cuenta.<br><br>" +
+        "¿Deseas continuar?</html>",
+        "Confirmar eliminación",
+        JOptionPane.YES_NO_OPTION,
+        JOptionPane.WARNING_MESSAGE
+    );
+
+    if (confirmacion == JOptionPane.YES_OPTION) {
         thebarbershop.utilidades.EliminarPerfil.eliminarPerfilCliente(this, emailUsuario);
+        // Nota: este método ya cierra la ventana y abre IniciarSesion
+    }
+        
     }//GEN-LAST:event_JBeliminarPerfilActionPerformed
 
     private void cargarDatosPerfil() {
